+++ keyboards/keychron/k8/keymaps/ansi/keymap.c
@@ -23,6 +23,21 @@
 // The underscores don't mean anything - you can have a layer called STUFF or any other name.
 // Layer names don't all need to be of the same length, obviously, and you can also skip them
 // entirely and just use numbers.
+enum layer_names {
+    WIN_BASE    = 0,
+    WIN_FN      = 1,
+    MAC_BASE    = 2,
+    MAC_FN      = 3,
+};
+
+#define KC_TASK LGUI(KC_TAB)        // Task viewer
+#define KC_FLXP LGUI(KC_E)          // Windows file explorer
+#define KC_CRTN LGUI(KC_C)          // Cortana | Microsoft Teams
+
+#define KC_MSSN LGUI(KC_F3)         // Mission Control
+#define KC_FIND LALT(LGUI(KC_SPC))  // Finder
+#define KC_SIRI LGUI(KC_SPC)        // Siri
+#define KC_MSNP LSFT(LGUI(KC_4))    // Mac snip tool
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 /*			+--------------------------------------------------------------------------+----------------+
@@ -92,6 +107,28 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
       _______, _______, _______,                   _______,                                     _______, _______, _______, _______,         RGB_SAD, _______, RGB_SAI
     )
 };
+
+bool dip_switch_update_user(uint8_t index, bool active){
+  switch(index){
+    case 0:
+      if(active) { // Mac mode
+          layer_move(MAC_BASE);
+      } else { // Windows mode
+          layer_move(WIN_BASE);
+      }
+      return false;
+    case 1:
+      if(active){ //Cable mode
+        // do stuff
+      }
+      else{ //BT mode
+        // do stuff
+      }
+      break;
+  }
+
+  return true;
+}
 
 void keyboard_post_init_user(void) {
   // Customise these values to desired behaviour
