+++ keyboards/keychron/k8/keymaps/via/keymap.c
@@ -73,14 +73,6 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 bool dip_switch_update_user(uint8_t index, bool active){
   switch(index){
     case 0:
-      if(active){ //BT mode
-        // do stuff
-      }
-      else{ //Cable mode
-        // do stuff
-      }
-      break;
-    case 1:
       if(active){ // Win/Android mode
         // do stuff
       }
@@ -88,6 +80,15 @@ bool dip_switch_update_user(uint8_t index, bool active){
         // do stuff
       }
       break;
+    case 1:
+      if(active){ //Cable mode
+        // do stuff
+      }
+      else{ //BT mode
+        // do stuff
+      }
+      break; 
+
   }
   return true;
 }
