+++ keyboards/keychron/k8/keymaps/via_iso/config.h
@@ -1,6 +1,5 @@
-/* Copyright 2020 Adam Honse <calcprogrammer1@gmail.com>
- * Copyright 2020 Dimitris Mantzouranis <d3xter93@gmail.com>
- * Copyright 2021 Harrison Chan (Xelus)
+/* Copyright 2021 Harrison Chan (Xelus)
+ * Copyright 2022 Thomas Bowman Mørch (tbowmo)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,5 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "k8.h"
+#pragma once
 
+// Enable RGB Matrix
+#define VIA_QMK_RGBLIGHT_ENABLE
