+++ keyboards/keychron/k8/rgb/ansi/config.h
@@ -21,3 +21,4 @@
 
 /* USB Device descriptor parameter */
 #define PRODUCT_ID      0xFE0E
+#define DEVICE_VER		0x0001
