+++ keyboards/keychron/k8/rgb/iso/config.h
@@ -21,3 +21,4 @@
 
 /* USB Device descriptor parameter */
 #define PRODUCT_ID      0xFE0F
+#define DEVICE_VER		0x0001
