+++ keyboards/keychron/k8/rgb/optical_ansi/config.h
@@ -19,8 +19,5 @@
 #define DRIVER_LED_TOTAL 87
 
 /* USB Device descriptor parameter */
-#define VENDOR_ID       0x3434
 #define PRODUCT_ID      0xFE0E
 #define DEVICE_VER		0x0001
-#define MANUFACTURER    Keychron
-#define PRODUCT         K8 RGB
